Move netlify config to top
I am not 100% sure this is correct, but I suppose netlify will look for this on the top level and it is irrelevant once deployed.
If this file is relevant once deployed we might have to put it into the `static` directly, because that's where assets that should be available once deployed go.